rtapi: Fix rtapi atomics to work with clang/clang++
