+++ src/rtapi/rtapi_uspace.hh
@@ -22,7 +22,7 @@
 #endif
 #include <unistd.h>
 #include <pthread.h>
-#include <atomic>
+#include "rtapi_atomic.h"
 
 static inline void rtapi_timespec_add(timespec &result, const timespec &ta, const timespec &tb) {
     result.tv_sec = ta.tv_sec + tb.tv_sec;
@@ -45,7 +45,7 @@ struct WithRoot
 {
     WithRoot();
     ~WithRoot();
-    static std::atomic<int> level;
+    static std::atomic_int level;
 };
 
 struct rtapi_task {
