+++ src/rtapi/uspace_rtapi_app.cc
@@ -51,14 +51,14 @@
 #include <pthread_np.h>
 #endif
 
+#include <boost/lockfree/queue.hpp>
+
 #include "rtapi.h"
 #include <hal.h>
 #include "hal/hal_priv.h"
 #include "rtapi_uspace.hh"
 
-#include <boost/lockfree/queue.hpp>
-
-std::atomic<int> WithRoot::level;
+std::atomic_int WithRoot::level;
 static uid_t euid, ruid;
 
 #include "rtapi/uspace_common.h"
